backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
PyJWT==2.9.0
email-validator>=2.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.9.0
pydantic-settings==2.5.0
python-multipart==0.0.9
python-socketio==5.11.0
